Done 2 cf +1lc, will do 1 more of each.

+++ spring.cpp
@@ -38,7 +38,7 @@ int main()
 
     while(t--)
     {
-        int a,b,c,m;
+        ll a,b,c,m;
         cin>>a>>b>>c>>m;
 
         /*
